pyproject.toml
[tool.poetry]
name = "code_count"
version = "0.1.0"
description = ""
authors = ["Michael Van Wickle <mvanwickle@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
attrs = "^19.3.0"
bokeh = "^2.1.1"
click = "^7.1.2"
funcy = "^1.14"
marshmallow = "^3.7.1"
matplotlib = "^3.2.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.2"

[tool.isort]
profile = "black"
no_lines_before = ["LOCALFOLDER"]

[tool.poetry.scripts]
ccount = "code_count.__main__:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
